Move song files (agent, buttons, loop) into src/songs/
