+++ package.json
@@ -13,8 +13,8 @@
     "random": "ts-node-dev --respawn src/random.ts",
     "index-trigger": "ts-node-dev --respawn src/index-trigger.ts",
     "preset": "ts-node src/preset-runner.ts",
-    "buttons": "ts-node-dev --respawn src/buttons.ts",
-    "agent": "ts-node-dev --respawn src/agent.ts",
+    "buttons": "ts-node-dev --respawn src/songs/buttons.ts",
+    "agent": "ts-node-dev --respawn src/songs/agent.ts",
     "control-server": "ts-node-dev --respawn src/control-server.ts"
   },
   "dependencies": {
